Updated with Zhou's comments
